trpe file moved to common
