--- src/components/common/dialogs/Dialog.type.ts
@@ -1,6 +1,0 @@
-export interface ForceDeleteDialogType {
-    onClickDelete: () => void
-    closeDeleteModal: () => void
-    forceDeleteDialogTitle: string
-    forceDeleteDialogMessage: string
-}
